Move parse_batch_slice back into batch_filter.py and reduce an import
